+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `#sub_accounts` to get the list of an accounts sub accounts
+
 ## [5.3.0] - 2022-09-09
 ### Added
 - `#sub_account_transfer_history`